+++ tests/testthat/test-gr_plot_sep.R
@@ -12,7 +12,7 @@ test_that("Period plot has the correct content", {
   # Four years in a matrix layout
   plt3 = suppressWarnings(gr_plot_sep(sep, 1988:1991, layout = matrix(1:4, nrow = 2, byrow = TRUE))) 
   # Four years in a matrix layout with free Y scale
-  plt3 = suppressWarnings(gr_plot_sep(sep, 1988:1991, layout = matrix(1:4, nrow = 2, byrow = TRUE), yfree = TRUE)) 
+  plt3 = suppressWarnings(gr_plot_sep(sep, 1988:1991, layout = matrix(1:4, nrow = 2, byrow = TRUE), yrange = 'each')) 
   # Add temperature
   plt4 = suppressWarnings(gr_plot_sep(sep, 1991, temp = TRUE))
   # Add precipitation